Refactor keldoc functions in different files
