+++ scraper/keldoc/keldoc_routes.py
@@ -1,0 +1,11 @@
+# Center info route
+API_KELDOC_CENTER = 'https://booking.keldoc.com/api/patients/v2/searches/resource'
+
+# Motive list route
+API_KELDOC_MOTIVES = 'https://booking.keldoc.com/api/patients/v2/clinics/{0}/specialties/{1}/cabinets/{2}/motive_categories'
+
+# Cabinet list route
+API_KELDOC_CABINETS = 'https://booking.keldoc.com/api/patients/v2/clinics/{0}/specialties/{1}/cabinets'
+
+# Calendar details route
+API_KELDOC_CALENDAR = 'https://www.keldoc.com/api/patients/v2/timetables/{0}'
