+++ scraper/prototype.py
@@ -8,7 +8,7 @@ import requests
 
 from .departements import to_departement_number, import_departements
 from .doctolib import fetch_slots as doctolib_fetch_slots
-from .keldoc import fetch_slots as keldoc_fetch_slots
+from .keldoc.keldoc import fetch_slots as keldoc_fetch_slots
 from .maiia import fetch_slots as maiia_fetch_slots
